+++ src/main/java/ar/edu/unq/desapp/grupoo/criptop2p/model/dto/IntentionDTO.java
@@ -3,13 +3,15 @@ package ar.edu.unq.desapp.grupoo.criptop2p.model.dto;
 import ar.edu.unq.desapp.grupoo.criptop2p.model.CryptoName;
 import ar.edu.unq.desapp.grupoo.criptop2p.model.Type;
 
+import java.math.BigDecimal;
+
 public class IntentionDTO {
     public int count;
-    public Long price;
+    public BigDecimal price;
 
     public Type type;
     public CryptoName cryptoName;
-    public IntentionDTO(int aCount, Long aPrice, Type aType, CryptoName aCryptoName) {
+    public IntentionDTO(int aCount, BigDecimal aPrice, Type aType, CryptoName aCryptoName) {
         this.count = aCount;
         this.price= aPrice;
         this.type = aType;
